fix migration

+++ db/migrate/20181016163502_add_price_to_perimeters.rb
@@ -1,5 +1,5 @@
 class AddPriceToPerimeters < ActiveRecord::Migration[5.1]
   def change
-    add_column :parking_perimeters, :price, :double
+    add_column :parking_perimeters, :price, :decimal
   end
 end
